+++ src/middlewares/src/use.ts
@@ -1,4 +1,4 @@
-import { MiddlewareCB } from '@heliosjs/core/types';
+import type { MiddlewareCB } from '@heliosjs/core/types';
 import { defineMiddlewaresMeta } from '@heliosjs/core/utils';
 
 /**
@@ -49,7 +49,7 @@ export function Use(middleware: MiddlewareCB | MiddlewareCB[]) {
   return function (target: any, propertyKey?: string, descriptor?: PropertyDescriptor) {
     const middlewares = Array.isArray(middleware) ? middleware : [middleware];
 
-    const data = middlewares.map(middleware => ({ middleware }));
+    const data = middlewares.map((middleware) => ({ middleware }));
 
     if (propertyKey) {
       defineMiddlewaresMeta(data, target, propertyKey);