+++ src/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 3.1.7
+
+### Patch Changes
+
+- Refactored controller prototype
+
 ## 3.1.6
 
 ### Patch Changes